refactor (enums) : 별도의 enums 패키지로 분리한다
